requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.3
Django==4.2.3
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.6
Flask==2.3.2
Flask-Cors==4.0.0
flask-restplus==0.13.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
MarkupSafe==2.1.3
packaging==23.1
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.7.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.6
uWSGI==2.0.24
Werkzeug==2.3.6
